Several changes
  - Worked to improve PID controller.  Should probably look to
    bring filter into Glow Worm so i can use it as a tool in the
    derivative.
  - Simple plant and control effort changed to use a long for
    u since it often exceeds 32000 when max RPM is around 10K.
  - Most importantly, quadrature was fixed to allows configuration
    for Uno or Due boards.

+++ blocks/simple_plant.h
@@ -80,7 +80,7 @@ public:
 	void run() {
 		//update the control effort
 		sub.update();
-		int effort = local_effort_msg.u;
+		long effort = local_effort_msg.u;
 		Direction::dir direction = (effort >= 0) ? Direction::fwd : Direction::bck;
 		effort = abs(effort);
 		effort = constrain(effort, 0, max_effort);
